+++ honeycomb/src/main/java/org/commonjava/o11yphant/honeycomb/interceptor/flat/FlatHoneycombWrapperEndInterceptor.java
@@ -16,7 +16,7 @@
 package org.commonjava.o11yphant.honeycomb.interceptor.flat;
 
 import io.honeycomb.beeline.tracing.Span;
-import org.commonjava.o11yphant.annotation.MetricWrapper;
+import org.commonjava.o11yphant.metrics.annotation.MetricWrapper;
 import org.commonjava.o11yphant.honeycomb.HoneycombManager;
 import org.commonjava.o11yphant.honeycomb.config.HoneycombConfiguration;
 import org.slf4j.Logger;
