+++ metrics/api/src/main/java/org/commonjava/o11yphant/metrics/annotation/MetricNamed.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.commonjava.o11yphant.annotation;
+package org.commonjava.o11yphant.metrics.annotation;
 
 import org.commonjava.o11yphant.metrics.MetricsConstants;
