+++ metrics/api/src/main/java/org/commonjava/o11yphant/metrics/annotation/MetricWrapperNamed.java
@@ -13,19 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.commonjava.o11yphant.annotation;
+package org.commonjava.o11yphant.metrics.annotation;
 
 import javax.interceptor.InterceptorBinding;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import static java.lang.annotation.ElementType.METHOD;
-import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @InterceptorBinding
-@Target( { METHOD, TYPE } )
+@Target( { PARAMETER } )
 @Retention( RUNTIME )
-public @interface MetricWrapperStart
+public @interface MetricWrapperNamed
 {
 }
