+++ metrics/api/src/main/java/org/commonjava/o11yphant/metrics/api/Gauge.java
@@ -1,4 +1,4 @@
-package org.commonjava.o11yphant.api;
+package org.commonjava.o11yphant.metrics.api;
 
 @FunctionalInterface
 public interface Gauge<T> extends Metric
