+++ metrics/api/src/main/java/org/commonjava/o11yphant/metrics/api/Meter.java
@@ -1,4 +1,4 @@
-package org.commonjava.o11yphant.api;
+package org.commonjava.o11yphant.metrics.api;
 
 public interface Meter
                 extends Metric, Metered
