+++ metrics/api/src/main/java/org/commonjava/o11yphant/metrics/api/Metered.java
@@ -1,4 +1,4 @@
-package org.commonjava.o11yphant.api;
+package org.commonjava.o11yphant.metrics.api;
 
 public interface Metered
 {
