+++ metrics/api/src/main/java/org/commonjava/o11yphant/metrics/api/Metric.java
@@ -1,0 +1,5 @@
+package org.commonjava.o11yphant.metrics.api;
+
+public interface Metric
+{
+}
