+++ metrics/api/src/main/java/org/commonjava/o11yphant/metrics/conf/MetricsConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.commonjava.o11yphant.conf;
+package org.commonjava.o11yphant.metrics.conf;
 
 public interface MetricsConfig
 {
