--- metrics/core/src/main/java/org/commonjava/o11yphant/api/Metric.java
@@ -1,5 +1,0 @@
-package org.commonjava.o11yphant.api;
-
-public interface Metric
-{
-}
