+++ metrics/core/src/main/java/org/commonjava/o11yphant/metrics/healthcheck/impl/HealthCheckResult.java
@@ -1,6 +1,6 @@
 package org.commonjava.o11yphant.metrics.healthcheck.impl;
 
-import org.commonjava.o11yphant.api.healthcheck.HealthCheck;
+import org.commonjava.o11yphant.metrics.api.healthcheck.HealthCheck;
 
 import java.util.Date;
 import java.util.HashMap;
