+++ metrics/core/src/main/java/org/commonjava/o11yphant/metrics/jaxrs/CodahaleHealthCheckDeploymentProvider.java
@@ -19,7 +19,7 @@ import com.codahale.metrics.servlets.HealthCheckServlet;
 import io.undertow.servlet.Servlets;
 import io.undertow.servlet.api.DeploymentInfo;
 import io.undertow.servlet.api.ServletInfo;
-import org.commonjava.o11yphant.conf.MetricsConfig;
+import org.commonjava.o11yphant.metrics.conf.MetricsConfig;
 import org.commonjava.o11yphant.metrics.DefaultMetricsManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
