+++ metrics/core/src/main/java/org/commonjava/o11yphant/metrics/system/SystemGaugesSetProvider.java
@@ -15,8 +15,8 @@
  */
 package org.commonjava.o11yphant.metrics.system;
 
-import org.commonjava.o11yphant.api.MetricRegistry;
-import org.commonjava.o11yphant.conf.MetricsConfig;
+import org.commonjava.o11yphant.metrics.api.MetricSet;
+import org.commonjava.o11yphant.metrics.conf.MetricsConfig;
 import org.commonjava.o11yphant.metrics.MetricSetProvider;
 
 import javax.enterprise.context.ApplicationScoped;
@@ -37,8 +37,20 @@ public class SystemGaugesSetProvider
     private static final String SYSTEM = "system";
 
     @Override
+    public MetricSet getMetricSet()
+    {
+        return systemGaugesSet;
+    }
+
+    @Override
+    public String getName()
+    {
+        return name( metricsConfig.getNodePrefix(), SYSTEM );
+    }
+
+    /*@Override
     public void registerMetricSet( MetricRegistry registry )
     {
         registry.register( name( metricsConfig.getNodePrefix(), SYSTEM ), systemGaugesSet );
-    }
+    }*/
 }
