+++ metrics/core/src/main/java/org/commonjava/o11yphant/metrics/util/NameUtils.java
@@ -15,7 +15,12 @@
  */
 package org.commonjava.o11yphant.metrics.util;
 
+import com.codahale.metrics.MetricRegistry;
 import org.apache.commons.lang3.ClassUtils;
+import org.commonjava.o11yphant.metrics.MetricsConstants;
+
+import static com.codahale.metrics.MetricRegistry.name;
+import static org.apache.commons.lang.StringUtils.isBlank;
 
 public class NameUtils
 {
@@ -64,4 +69,46 @@ public class NameUtils
 
     }
 
+    /**
+     * Get default metric name. Use abbreviated package name, e.g., foo.bar.ClassA.methodB -> f.b.ClassA.methodB
+     */
+    public static String getDefaultName( Class<?> declaringClass, String method )
+    {
+        // minimum len 1 shortens the package name and keeps class name
+        String cls = ClassUtils.getAbbreviatedName( declaringClass.getName(), 1 );
+        return MetricRegistry.name( cls, method );
+    }
+
+    /**
+     * Get default metric name. Use abbreviated package name, e.g., foo.bar.ClassA.methodB -> f.b.ClassA.methodB
+     */
+    public static String getDefaultName( String declaringClass, String method )
+    {
+        // minimum len 1 shortens the package name and keeps class name
+        String cls = ClassUtils.getAbbreviatedName( declaringClass, 1 );
+        return MetricRegistry.name( cls, method );
+    }
+
+    /**
+     * Get the metric fullname with no default value.
+     * @param nameParts user specified name parts
+     */
+    public static String getSupername( String nodePrefix, String... nameParts )
+    {
+        return MetricRegistry.name( nodePrefix, nameParts );
+    }
+
+    /**
+     * Get the metric fullname.
+     * @param name user specified name
+     * @param defaultName 'class name + method name', not null.
+     */
+    public static String getName( String nodePrefix, String name, String defaultName, String... suffix )
+    {
+        if ( isBlank( name ) || name.equals( MetricsConstants.DEFAULT ) )
+        {
+            name = defaultName;
+        }
+        return MetricRegistry.name( MetricRegistry.name( nodePrefix, name ), suffix );
+    }
 }
