+++ metrics/pom.xml
@@ -31,6 +31,7 @@
 
   <modules>
     <module>core</module>
+    <module>api</module>
   </modules>
 
 </project>
