+++ pom.xml
@@ -84,6 +84,11 @@
             </dependency>
             <dependency>
                 <groupId>org.commonjava.util</groupId>
+                <artifactId>o11yphant-metrics-api</artifactId>
+                <version>1.0-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>org.commonjava.util</groupId>
                 <artifactId>o11yphant-metrics-core</artifactId>
                 <version>1.0-SNAPSHOT</version>
             </dependency>
